Fix tspacket by indexing off the payload rather than header
The actual AFCOffset() is indexed off the beginning of the packet, however, the
StartOfFieldPointer() is now indexing off the _payload[] rather than the _data[]
which is not in TSPacket, but rather in TSHeader (which TSPacket inherits from).

Although the cast done from char * to TSPacket * puts the data in the correct
place, the compiler is not at all happy with it when you run off the end of
the _data[] array and into the _payload[] array.  Since that is what we really
are intending to do, let's just do it explicitly and make the compiler happy
in the process.