Add a new DVDInfo class.
This is a simple DVD helper that simply opens the device and retrieves
the name and serial number of the disc. A complete DVD device, as
implemented in DVDRingBuffer, can be expensive to fully create and open
and when performing the initial DVD bookmark check, we need nothing more
than the basics.

For remote, storage group based playback this can knock up to 5 seconds
off the time taken to start a disc.