Improved DVD startup.
- probe the dvd titles to identify the first title that is definitely a
video sequence (as opposed to a still frame) and use this for the
initial startup/probe.

- retrieving chapter data is expensive over storage groups (potentially
2-3 seconds per title for up to 70 titles or more), so don't necessarily
scan all of the titles but do cache the results (they shouldn't change).

- the chapter data will also be used for improved chapter navigation.

This fixes/improves a few startup issues. By picking a valid video
sequence, we avoid some video artefacts at the beginning of playback and
avoid having to reset certain discs that have hit a stop sequence
(because we've scanned a menu) - with a knock-on improvement in startup
speed.