Refactor DVD chapter navigation.
- move chapter/track seeking out of the decoder class and into
DVDRingBuffer.

- reimplement the MythPlayer chapter handling methods in MythDVDPlayer
for a consistent approach.

- add a new method to DVDRingBuffer to allow jumping to a specific
chapter.

All told, dvd chapter navigation is now consistent between matroska
(i.e.ripped) versions of a title and the original dvd version.

As noted previously, retrieving chapter data for a title can be
expensive and blocks the UI, hence there may be a short pause when
displaying chapter times for a title that is not aleady cached. The
chapter data for the main title is usually, but not always, retrieved
during the initial dvd scan.