PiP: Remove the extra thread created for PiP playback.
This is no longer needed as all players are run from the main thread. By
and large, the 'playback loop' for a pip player is a no-op as the
decoding is elsewhere and the actual processing of video frames is
handled by the main player (though a small optimisation is needed for
the VideoSync).