Split the database logging off into another thread
To deal with logging to the database when the database is not available, I
have added a queue for use by the database logger, and given the database
logger its own thread.  One complication of this approach is that a reference
counter had to be added to logging items so they can be cleanly deleted when
done logging.