Don't write to DB log unless logging table exists.
Verify the logging table exists before writing to the DB log.  This will
only really come into play on new databases and the initial upgrade to
the new-logging DB schema.  While the approach isn't necessarily the
most elegant, it's definitely the simplest, and shouldn't be any
strain at all on any system with a new-logging-capable schema (and is
only a minor strain on a system updating to such a schema).