UPnP: Fix SSDP Notify messages.
At some point in the dim and distant past, the socket used for SSDP
alive/byebye messaging was converted to QMulticastSocket. The conversion
however did not pick up the use of the member variables from the new
subclass and hence SSDP notify messages are typically being sent to
0.0.0.0.

The fix just ensures we use the QMulticastSocket object properly. There
may be a slightly simpler fix to re-use the address and port in the
MSocketDevice parent class but I have no idea if that would be safe.