MythFrontend status: Add details on slave backends and other frontends.
- perform full SSDP discovery when starting the frontend and UPnP is
enabled. This ensures the frontend knows of all other UPnP devices on
the network (currently it will notify all existing devices of its
existance when starting but will not be aware of those devices).

- add information on other known frontends and slave backends to the
staus page. As noted, the filtering out of the current frontend may not
always work (but is harmless) and only one frontend will ever be
displayed per IP address (currently, UPnP will fail to start after the
fist frontend as it fails to bind to the desired port).