This patch adds IPv6 support to mythbackend and mythfrontend. I've been running this patch since XMas with no problems at all.
To use IPv6 you just have to run mythtv-setup and configure your backend IPv6 addresses.
This assumes  you have IPv6 configured internally on your lan of course.

The changes are overall fairly small.  One of the biggest things needed
was to encapsulate IPv6 addresses in the URL with [] (otherwise parsing
out a port could be ugly:) )

So for example instead of

   myth://group@fd14:6bf7:21fb::24/file.mpg

 It would be
   myth://group@[fd14:6bf7:21fb::24]/file.mpg

   or if specifying the port

   myth://group@[fd14:6bf7:21fb::24]:6543/file.mpg

To remove all of the duplicated code where we build
the URL I've added a GenMythURL() function to gContext and it will generate
the path and do the right thing when it comes to[]'s , ports  and storage
groups in the URL.

So for example gContext->GenMythURL("fd14:6bf7:21fb::24","6543","file.mpg","Videos")
Happily generates myth://Videos@[fd14:6bf7:21fb::24]:6543/file.mpg

Only the PHP bindings have been updated (for use with mythweb) so far and I have no had
a chance to test this under Windows or OSX as I do not have access to either of those.

This patch does however change MythSocket to no longer assume that a socket will be IPv4.
It will attempt to detect which it should use however based on the IP it is trying to connect to

Bug reports are welcome.