Redo the ffmpeg thread init code
The warnings have been there long enough.  Seems that with the current
libavcodec, you just need to set the thread_count before opening the avcodec,
and it will actually start the threads with no other intervention, and close
them down on the avcodec_close.  This should get rid of those warnings.

If this causes an issue with framegrabber recordings, please let me know.