MythDownloadManager: expose the cookieJar used by the download manager.
The main reason to do this is so MythUIWebBrowser can use the same cookiejar as
the download manager. It would be nice to share the same QNetworkAccessManager
but that is not possible because the download manager runs in its own thread so
the next best thing is to use a shared cookiejar which keeps some sites like
googlemail for example happy when trying to download a file.