Begin updating strings in HTTP setup.
This patch is the first in a series of patches that will update the
strings in backend HTTP setup to agree with the changes we have made to
the strings in programs/mythtv-setup/backendsettings.cpp since the
original XML files were created (over a year before the XML files were
added to master).  This should make more of the translations work (and
prevent translators from having to re-tranlate strings that were
supposed to have been removed and replaced with better strings).

I have made every attempt to not revert changes that were made only in
the XML files with the intention of clarifying the strings (and may
eventually backport them to mythtv-setup's backend settings just so we
have a single string for translators).  Robert K has independently made
a couple of the changes that were already made to mythtv-setup's
strings, but this series (when complete) should prevent further
duplication of efforts.

I've only checked settings up through UPNP Server Settings in
config_backend_general.xml, but wanted to get this in while it still
applies, since I likely won't work on this further until next week.

Changes incorporated as part of this series (will) include those made in
7c46b759357239423ff144203e2f9492a14dfb0c ,
75dc7cf1711f755fb02fe8572b9ce2b333f42f0a,
24e854a55c21f8ecf4f1e2ff483089f74674e9a7,
fe1c8bff491e947a990b4e88a16493cac49bc028,
fcfe99683b014aba92055696c3062a0726e1d6eb , and possibly others

In addition, this patch:
 - Adds StorageScheduler option BalancedPercFreeSpace (from
   b9efe922d71dc4c19e7ec852b925f5dea7c6ff16 )
 - Removes EITIgnoresSource setting (from
   c2c7de48115b6bcdfdf5ecbb661f427705362374 )