Allow for drivers that do not use VIDIOC_ENCODER_CMD ioctl
Not all V4L2 drivers support this yet (marked experimental in the V4L2 API
specification).  Failing hard on a driver that doesn't support it (or need it)
is probably not our best bet.

This shouldn't be an issue on the StopEncoding, as if we fail hard while trying
to stop encoding, we still get the desired effect, although we would not be
stopping at a GOP boundary.