Change the custom aspect ratio code
The custom aspect ratio calculations were not always making it to the db
correctly.  Too many ways of getting the aspect ratios, I guess.  Anyways, the
only thing that seems to even use this is the flowplayer in mythweb, and a
corresponding change to that code will be committed shortly.