Initialize the MythUIImage filename correctly
Fixes #9634.  Fixes #9708.  Refs #9547.

If the filename does not start with "/", "http://", "https://" or "ftp://", then
the actual filename loaded is uninitialized.  It can be any random crap on the
stack.  I'm honestly surprised this worked at all :)

This may fix things for #9547 as well, and should be tested with that
scenario too.