Silence ffmpeg log spew during av_find_stream_info
As we start playback, there is often a lot of spew while ffmpeg libs determine
the stream info, particularly for H.264 recordings.  This is even the case when
there is no playback issues at all with the file.

What I've done is added a flag to totally silence the ffmpeg log spew, but only
activate it around the call to av_find_stream_info in
AvFormatDecoder::FindStreamInfo method.

This will remove a pile of useless log spew that really doesn't help us in
any debugging.