MythUI: enhance <minsize> functionality.
Allow <minsize> areas to be siblings of each other, with each contributing
to the resulting area.

Allow <minsize> areas to be nested, with the child areas impacting the
resulting area.

Add a <initiator> element so a theme can control when <minsize> areas
instigate a resize, and which <minsize> areas just react to the resize.  By
default <textarea> elements are initiators, and all others are not.  A patch
to allow <imagetype> elements to be initators is forthcoming.