Try to improve row counting performance for logging.
Since InnoDB engine doesn't store the row count for tables, like MyISAM
does, try to improve the performance of the count query.  By counting
id, the optimizer should choose to use the primary index.  At this
point, I'm not forcing it, but if it's found to help for InnoDB users,
we can do so later.

See:
http://www.cloudspace.com/blog/2009/08/06/fast-mysql-innodb-count-really-fast/
http://www.mysqlperformanceblog.com/2007/04/10/count-vs-countcol/