libmythupnp: New UPNPSubscription extension.
An HttpServer Extension that manages subscriptions to UPnP services.

An object wishing to subscribe to a service needs to register as a
listener for events and subscribe using a valid usn and subscription
path. The subscriber is responsible for requesting a renewal before the
subscription expires, removing any stale subscriptions, unsubsubscribing
on exit and must re-implement QObject::customEvent to receive event
notifications for subscribed services.

Tested with gmediastreamer, MediaTomb, Play On and Windows Media Player
11.