Make mythtranscode fail gracefully on file not found
Fixes #9427

If a recording doesn't exist, or the recording exists, but is not in the
database, mythtranscode will gracefully refuse to transcode it, rather than
crash.

Also, make the MythContext object be on the stack of main(), such that any exit
from the system will hit the dtor, causing a cleaner shutdown.