Blu-ray Playback: Upstream updates and use newly added API.
Adds a bd_read_skip_still() function.  Basically this needs to be called at the end of a still frame (which we really don't support yet) before bd_read() will move on to the next title.  So, for now, just put it in the if statement where we print verbose about not supporting still frames.

All in all, no real change to the behavior, but needed to implement this to continue to keep up with updates.