RingBuffer: Add bitrate monitoring code.
- enable monitoring with EnableBitrateMonitor(true)
- poll for data via GetDecoderRate, GetStorageRate, GetAvailableBuffer
- disable monitoring with EnableBitrateMonitor(false)

The decoder rate is the rate at which the decoder has requested data
from the buffer over the previous second. The storage rate is the
average rate at which data has been delivered to the buffer in the
previous second. The available buffer is the % of the buffer that is
available for the decoder.

N.B. the storage rate can be misleading. For dvd and bluray playback the
rate shown is the speed of retrieving data from the libdvdnav/libbluray
object - there is another ringbuffer layer that performs the actual
reads from local or remote storage. For network mounted drives and
http:/ streaming, there is additional buffering that may suggest that
read speeds are in excess of the actual network capabilities.