Remove O_SYNC from the debug log files
This was a good suggestion on the part of Michael T. Dean.  There is no
particular reason to do synchronous writes to the debug log files, and doing so
actually causes the logger to fall behind under high load.  With our writes
going to the file asynchronously, depending on the OS buffering, logging
keeps up much better.