RingBuffer: Make the buffer size dynamic.
- convert the buffer size to a member variable that can be adjusted mid-
- add buffer sizing factors that increase the size of the buffer for
networked playback, matroska streams and streams with an unknown video
bitrate.
- set the matroska and unknown bitrate flags when scanning the streams
in AvFormatDecoder.

RingBuffer behaviour is unchanged for playback of local files and for
RingBuffers operating on backends; the default 4Mb is ample for fast,
local disk access.

For networked playback, the 4Mb is doubled and doubled again for both
matroska files and files of an unknown bitrate - and hence the maximum
networked buffer size (in the frontend) is 32Mb. This is sufficient to
playback the now infamous galactica clip over an 18Mps wifi connection
with minimal pauses (and there are a couple of additional tweaks that
will remove those pauses altogether).

The specific triggers and multipliers used are a little contrived to fix
the worst case scenarios we have now (in many cases the buffer size will
be much larger than needed - improvements are more than welcome) but the
increase in buffer size for networked playback is probably long overdue.