Add groundwork for enhanced recording rule filters.
This change adds some of the capabilities of custom recording rules to
normal rules.  This is very convenient as the desire to use common
filters no longers requires having to create custom rules.  In
addition, the filter scheme is extensible so that new filters can be
easily added without code changes.

IMPORTANT NOTE: There is no UI support for the new filters in this
commit.  Someone else has graciously volunteered to help with that and
it will be added in a future commit.  Adding the core and database
changes now makes cooperating on the UI work easier.

IMPORTANT NOTE 2: The 6 filters included in this change are intended
for testing purposes only at this time.  The actual filters and
associated ids are subject to change before the final commit for this
feature.  IOW, don't use them for anything important unless you are
prepared to manually fix your recording rules later.