MythMusic: change the order of doing things in loadMusic()
This fixes a problem with both the load music and file scanner progress dialogs
both being shown at the same time. This can happen on new installs if the user
does something to trigger a loadMusic() and the file scanner is started
automatically because there is no track data in the database.