libmythdvdnav: Resync dvdnav and dvdread.
This covers upstream SVN revisions 1216-1226.  Covers the following changes:

libdvdnav:

* Prevent abort if a menu doesn't exist.
* Fix typos
* Fix grammar in dvdnav.h
* Fix typos in comments

libdvdread:

* Provide BUP file support for more issues.
* Mark a large number of functions not used outside of the file as static.
* Fix some spelling issues
* Move UDFReadBlocksRaw declaration to dvdread_internal.h.
* Inhibit double free of parental management information
* Replaces the hard-coded values (in parental management)
* Report which index failed reading

Fixes #9828.  The ticket contained a patch which purports to fix additional memory leaks in the upstream copy, but I'm going to insist that that patch be applied upstream first for my sanity's sake.  Given the slow rate of change in libdvdnav and libdvdread, I am willing to perform a resync when pinged on this after it's committed there.

Tested this resync with local, disc, and myth:// playback of ISOs and DVDs.