RingBuffer: Tweak the fill_threshold settings again.
- raise fill_threshold to 7/8th of the buffer size, which gives up to
4Mb of additional filled buffer.
- drop wait timeout for filling the additional buffer space above the
threshold from 1000ms to 50ms.

On bursty matroska streams, there may be no buffer activity at all for
10s of seconds. During this time the read ahead loop would only trigger
an additional read above the fill threshold every second and under the
worst case scenario, this meant that the buffer was not sufficiently
full before the next burst of read activity.

(N.B. I inadvertantly committed a change to the fill threshold in
bff85941ca)