Fix crash on getting filesystem usage
I'm not sure why this didn't crash before the recent FilySystemInfo changes,
but for sure it does with.  We were reading the wrong number of iterators,
causing the toInt() to be called on non-ints, causing fun.  Also, on the last
loop, we'd be reading from iterators not in the list as we were out of whack.