libmythui: Fix a leak (flood) in MythPainter.
The software cache size was never increased as the images hadn't been
created when we queried their size - and hence nothing was ever removed
from the software cache.

There will still be reports of un-allocated images in the MythPainter
destructor. These are the main OSD window images that are still being
held by the main UI image cache. Not yet sure of the best approach with
those - I don't really want to leave them hanging.

Refs #9707.