Add a vanish attribute to minsize
If vanish is enabled, and the "contents" of the area are empty, then the
whole area will disappear.

For example, if a textarea and it's sibling shape both have vanish set, then
if the "text" in the textarea is empty, both the textarea and the shape will
vanish -- they won't be drawn, and won't take up any space in the button.