Merge MythVideo into core.
Most definitely requires a make distclean of both core and plugins.  To be safe, you should also probably remove the libmythvideo.so file from your installed plusings dir ($PREFIX/libs/mythtv/plugins).

Move all the UI portions of mythvideo into the core of MythTV.  The concept of a TV-only media center has become increasingly irrelevant, and this is the first move towards bringing obvious and necessary functionality in to the core.

Part of this work was already done before .24 with the creation of libmythmetadata, which held all the non-UI portions of mythvideo.  Since the DB changes were committed to make sure that all users now have a videometadata table, the only thing that remained was to bring the UI itself into mythfrontend.  This work is now more or less done.  There's a lot of interconnected menu themes, bits, and pieces, but my testing of all menu themes and all obvious video settings, views, and popups seems to indicate that all is working properly.  I'm sure that at least *something* has escaped me, but we'll handle those issues as they arrive.