ChannelScan: Update multiplex transponder and network id when channel updated
When a transport is added manually using the transport editor, the transport
and network ids are not known and are left unset.  This causes a problem
for interactive TV which tunes a channel by network and transport id.

This patch updates the transport and network ids for a multiplex whenever a
channel is found or updated on the transport.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>