Change cout to logging where appropriate
Also some formatting (80-col wrapping) changes, and changing commented code
blocks to #if 0/#endif to allow for nested comments better.

Changed the debugging of scheduling conflicts to be able to be enabled by the
existance of the DEBUG_CONFLICTS environment variable.  The previous code
required a recompile to enable it, but still had if() blocks around the debug
code.  It was either this or convert to #ifdef DEBUG_CONFLICTS.