Convert many cerr to VERBOSE
As many of these are debugging that really could and should be in the logs
(especially if we want to see them in user logs rather than have them use
"mythbackend 2>file")...  I have converted a pile of cerr uses to VERBOSE.

I have purposely skipped mythtranscode and mythcommflag for the moment as both
of those need the console output worked on anyways now that logging is on their
consoles as well.  I don't have the energy tonight to determine a sane way to
deal with it.  Also skipped are any that seemed to be bona-fide CLI output.

Also did some 80col reformatting where I was already in there.

Of course, all of this can be tweaked, and a pile of it is commented code
anyways.