OSD: Fix subtitle window painting order.
The order of iterating over a QHash is arbitrary and the fact that
subtitle and interactive screens have (mostly) been drawn in the correct
order since the OSD re-write is pure luck.

So switch to a QMap which, while marginally slower, will order on the
basis of the key (QString) and force the subtitle/interactive screens to
be painted first by adding some numerics to their identifiers.

Should fix the ordering issue seen in #9822 and occasionally in the yet
to be enabled blu-ray HDMV navigation mode.

Refs #9822