DVB Subtitles: Fix display of subs with no display segment.
The DVB subtitle standard reads:

'Within a display set the sequence of segments (when present) is:
- page composition;
- region composition;
- CLUT definition;
- object data;
- end of display segment;'

In the samples provided in #9373, the end of display segment is not
seen. The 'when present' above suggests that it is not required but the
ffmpeg code assumes it is. The workaround is to check that the previous
4 segments have been seen and, if so, process a dummy display segment.

mplayer and ffplay both appear to have the same problem.

Cherry-picked from master d27bedf5571bdb80c164

Closes #9373