Channel Scanner: Try harder to generate unique callsigns.
For some users of our channel scanner/some locales, we try to get the shortname from the SDT.  So far, so good.  Unfortunately, when this isn't available, the callsign can end up empty or non-unique, which messes up scheduling for some of those users.

Instead, generate a callsign based on the TSID and ServiceID, which should fairly reliably create unique callsigns.  They won't be the world's most informative callsigns, but scheduling and guide browsing should work as expected, which is important.