DVDRingBuffer: Bump up the minimum title length for initialisation.
The DVD used in #9170 has a still frame with a duration of 10 seconds as
its first title. Playback fails as we cannot test for stillframe status
until we've started playing the title, ffmpeg is unable to find enough
data to create a decoder and the DVD object refuses to read beyond that
still frame.

The quick fix here is to increase the minimum duration of the initial
title selected for probing - it is highly unlikely that any still frame
is on screen for more than 60 seconds (an infinite still frame is
reported with a small duration) AND is not followed by a video segment.

A better fix would be to check the still frame status and try and find
another title if necessary but I cannot find a disc that has the
necessary structure to test with.

Refs #9170