MythFrontend: Add a new Slave menu theme.
A menu theme with no menus. It is intended for frontends that are
operated remotely via either the network or http interfaces.

It is not selectable from the appearance settings (as you can't change
it back!) and can currently only be used in a non-persistent manner from
the command line with:

mythfrontend -O MenuTheme=Slave

In the future there may be a more convenient and obvious command line
switch and 'slave' frontends may advertise themselves as such through
the UPnP server (in much the same way as slave backends distinguish
themselves from master backends).