Rewrite of CoreAudio driver.
While not everything is fixed yet, it is a step in the right direction.

-You can now select which device to use in the setup and not always have to use the default audio device.
-Fixes #5552.
-Properly set channel layout, under most circumstances, you shouldn't have to use Audio Midi Setup utilities

There are still bugs that I can't explain, probably due to my lack of deep knowledge of CoreAudio, and it certainly ain't behaving as it should. In particular, switching from AC3/DTS bitstream back to PCM will fail except if you have selected the default output device.
Until this is resolved, it is best to keep using the default output device.