More aggressively move recordings to resolve conflicts.
Change the scheduler to move already scheduled programs in order to
resolve conflicts as long as the recording isn't moved to a lower
priority showing.  Previously, the scheduler would only do this in one
very restictive case.  N.B. the scheduler might still move a recording
to a lower priority showing in a later pass if the "Reschedule higher
priorities" setting is enabled.