VideoOutputOpenGL: Use a shared OpenGL context for OpenGL video and UI.
This should fix:-

- blank OpenGL video playback on OS X.
- framebuffer 'confusion' where the driver gets stuck displaying a UI
context backbuffer during video playback and vica-versa. Seen on intel
clarkdale and AMD with the proprietary driver.

This also clears the way for VAAPI integration and the removal of
context locking and switching code that is reducing performance and
breaking OpenGL rendering on windows.

In the medium term, integrated video and UI rendering is now
straightforward.

Refs #9268
Refs #8593