Indicate log message level in console and debug logs
Right after the timestamp, I added a single character that maps to the log
message's level.  This will not be too useful until I have completed conversion
of the VERBOSE -> LogPrint that I have planned.  Currently, all VERBOSE maps
to LOG_INFO, so you will see a lot of "I".

For LOG_EMERG, I used "!" as "E" was already used for LOG_ERROR.