Add tid to the debug log file
Instead of [pid], it will now show [pid/tid] where tid is the hidden pid of the
underlying thread.  This matches the thread number shown in gdb backtraces:

e.g.
Thread 21 (Thread 29810):

In that example, 29810 would be the tid.

This will only work for Linux at this time as it uses the gettid system call,
which I have only found for Linux so far.

The tid is also logged in register/deregister debug log messages.