Use the current video aspect ratio if the video stream is reset.
In AvFormatDecoder, if a stream change is detected and no
new/replacement video stream is found, we choose some sensible defaults
and tell the player about the change. The video aspect ratio is however
reset to the default 4/3 - which immediately changes the currently
displayed frames. Instead use a pre-existing aspect ratio, which will be
4/3 by default if we haven't yet determined something more sensible.

This is only likely to be an issue with DVD playback, where we often get
stream updates piecemeal from the demuxer - the streams are reset and
some new audio streams added (no video found) and the new video stream
is added a 'short' period of time later.

Refs #9854