Run the image last modified check in the background for non-local images.
When loading an image, we normally check the last modified date to
determine if we can use the cached image or if we need to load a new
copy.  For non-local images (myth://, http://, etc.), this check can
be expensive, so perform all non-local last modified checks in the
background in the image loading thread.  This should speed up the UI
drawing when there are non-local images on screen.