mythtv-setup: Fix scrolling in the xmltv config terminal screen.
This allows you to both scroll up or down in the terminal and to escape out
of it. It also always makes sure the last line of the terminal output is
visible. Fixes #7962.

This is supposed to be fixed in the settings re-write but since it's been
delayed to after 0.25 let's at least make it usable in the mean time.
(cherry picked from commit 9861cf29b6fd5ca49ae573dfa7a5ff795f2bfdb5)