MythUIButtonTree: Fix an abort when scrolling.
The sanity checks in SetTreeState() were wrong causing an occasional abort when
scrolling around. This actually fixes the crash in ticket #9620 but the root
cause still remains which is the wrong node somehow gets the focus after
the tree is assigned if the top level node has two or more nodes. Refs #9620.