Reimplemented the VB_* definitions
The old macros made for confusing code that was rather hard to maintain.  I
changed it to store the verbose mask options in a QMap, which also has the
advantage of far simpler searching for the options while parsing.

Additionally, I have attempted to expand the size of the mask to 64 bits,
although we will have to see if it likes extending enum past 32 bits.  All I've
read online seems to indicate that enum is designed to auto-size and that you
shouldn't depend on it always being 32-bit, so hopefully when assigned a value
> 32bit in size, gcc will do the right thing.  The internal variables have been
changed to uint64_t in any case, so if there's borkage, I expect it will be in
creating the enum itself (which is in verbosedefs.h now)