Added support for setting the minimum log level
As we have several logging choices now, and the logging will be split to use
a mask (VB_*) and a log level(LOG_*), we need a way to specify a minimum log
level to care about.  Any messages at a lower priority level will be summarily
dropped.  The default is the info level (LOG_INFO).

This is in preparation for the splitting of VERBOSE -> LogPrint with separate
mask and level defined for each log message.