Merge mythverbose into mythlogging
Since there is a large overlap in functionality, it's about time to merge these
together.  I wanted to put in the propagation of verbose mask and log level
but they were in two separate files.

Note: although this does hit a LOT of files 90% of this is changing the name
of the included header file.

Also changed: print_verbose_messages -> verboseMask,
parse_verbose_arg -> verboseParseArg, safe_eno_to_string -> logStrerror